# docs/swagger.yaml
definitions:
  curd.ReplyData-types_User:
    properties:
      data:
        $ref: '#/definitions/types.User'
      error:
        type: string
    type: object
  curd.ReplyList-types_User:
    properties:
      data:
        items:
          $ref: '#/definitions/types.User'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  types.User:
    properties:
      admin:
        type: boolean
      cellPhone:
        type: string
      create:
        type: string
      disable:
        type: boolean
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /user/{id}:
    post:
      consumes:
      - application/json
      description: 修改用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_User'
      summary: 修改用户
      tags:
      - user
  /user/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_User'
      summary: 禁用用户
      tags:
      - user
  /user/create:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 用户信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_User'
      summary: 创建用户
      tags:
      - user
  /user/list:
    get:
      consumes:
      - application/json
      description: 查询用户
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_User'
      summary: 查询用户
      tags:
      - user
swagger: "2.0"